fixed everything cuz im ranson w/ jennys help :\
