--- app/server/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "server",
-  "version": "0.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "BSD-2-Clause",
-  "dependencies": {
-    "express": "~4.10.6",
-    "mongoose": "~3.8.21",
-    "jade": "~1.8.2"
-  },
-  "devDependencies": {
-    "body-parser": "~1.10.0"
-  }
-}
